auth-service/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
asyncpg==0.29.0
bcrypt==4.2.0
black==24.8.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
jwt==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.1.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.0
Werkzeug==3.0.4
